moved DesiredSkillsPicker to UserProfile/components/
