+++ src/components/UserProfile/components/DesiredSkillsPicker/DesiredSkillPicker.scss
@@ -1,4 +1,4 @@
-@import '../../styles/abstracts/_variables.scss';
+@import '../../../../styles/abstracts/_variables.scss';
 
 .chosen-skills {
   @include portal-description;
