+++ src/components/UserProfile/components/index.jsx
@@ -1,0 +1,7 @@
+import DesiredSkillsPicker from "./DesiredSkillsPicker";
+import ChosenSkills from "./ChosenSkills";
+
+export {
+  DesiredSkillsPicker,
+  ChosenSkills,
+};
